Unify migrators
